add_libft

+++ ft_bzero.c
@@ -1,6 +1,8 @@
-void ft_bzero(void *s, int n)
+#include "libft.h"
+
+void ft_bzero(void *s, size_t n)
 {
-    int i;
+    size_t i;
 
     if(!s)
         return;
